backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.31.1
httpx>=0.27.1
beautifulsoup4>=4.12.3
sqlalchemy>=2.0.36
apscheduler==3.10.4
python-dotenv>=1.0.1
